staticdata: stop compressing and storing large IR
Compressing IR adds Methods roots, which tend to become quite expensive
to store after some time. We also used to store the inferred code, but
since we never actually look at that again now (since now the compiler
avoids using --output-ji or --sysimage-native-code=no or --pkgimages=no)
it is now quite a bit of wasted space also.

For some reason, this also generates 20% more native code in the system
image, which isn't quite clear if that is good or not.